+++ lib/screens/home/top/categories/categories.dart
@@ -3,8 +3,8 @@ import 'package:flutter/services.dart';
 import 'package:flutter_mobx/flutter_mobx.dart';
 import 'package:frosty/api/twitch_api.dart';
 import 'package:frosty/core/auth/auth_store.dart';
-import 'package:frosty/screens/home/stores/categories_store.dart';
-import 'package:frosty/screens/home/widgets/category_card.dart';
+import 'package:frosty/screens/home/top/categories/categories_store.dart';
+import 'package:frosty/screens/home/top/categories/category_card.dart';
 import 'package:frosty/widgets/alert_message.dart';
 import 'package:frosty/widgets/loading_indicator.dart';
 import 'package:provider/provider.dart';
