+++ lib/screens/home/top/top.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:frosty/screens/home/stores/home_store.dart';
-import 'package:frosty/screens/home/stores/list_store.dart';
-import 'package:frosty/screens/home/top/categories.dart';
-import 'package:frosty/screens/home/widgets/streams_list.dart';
+import 'package:frosty/screens/home/home_store.dart';
+import 'package:frosty/screens/home/stream_list/stream_list_store.dart';
+import 'package:frosty/screens/home/stream_list/streams_list.dart';
+import 'package:frosty/screens/home/top/categories/categories.dart';
 
 class TopSection extends StatefulWidget {
   final HomeStore homeStore;
